scripts/embarc_tools/requirements.txt
PyYAML
colorama
PrettyTable
Jinja